Controller can run

+++ api/v1/hellojob_types.go
@@ -31,7 +31,7 @@ type HelloJobSpec struct {
 	// Foo is an example field of HelloJob. Edit hellojob_types.go to remove/update
 	Image        string `json:"image"`
 	Message      string `json:"message,omitempty"`
-	DelaySeconds *int64 `json:"delayseconds,omitempty"`
+	DelaySeconds *int   `json:"delayseconds,omitempty"`
 }
 
 // HelloJobStatus defines the observed state of HelloJob
